rentfree/requirements.txt
--no-binary=wagtail-2fa
bleach==3.3.1
beautifulsoup4>=4.8,<4.10
python-dotenv>=0.19.2
Django>=3.2.12,<3.3
phonenumberslite>=8.12.30
django-allauth>=0.48.0
django-allauth-2fa>=0.8
django-anymail>=8.4
django-bootstrap5>=21
django-phonenumber-field[phonenumberslite]>=5.2
django-storages>=1.11.1
django-comments-dab>=2.7.1,<2.7.2
dj-stripe>=2.6.1
boto3>=1.20.51
lxml>=4.7.1
psycopg2>=2.9.3
git+git://github.com/rentfreemedia/django-betterforms@changes
git+git://github.com/rentfreemedia/django-post_office.git@changes
git+git://github.com/rentfreemedia/wagtail-cache.git@changes
git+git://github.com/rentfreemedia/python-video-ids.git@changes
git+git://github.com/rentfreemedia/django-summernote.git@changes
git+git://github.com/rentfreemedia/django-drip-campaigns.git@changes
git+git://github.com/rentfreemedia/wagtail-personalisation.git@changes
git+git://github.com/rentfreemedia/django-dbtemplates.git@changes
git+git://github.com/rentfreemedia/markdown.git@changes
pytube>=11.0.1
unidecode>=1.3.2
wagtail>=2.15,<2.16
wagtail-2fa>=1.5.0
wagtailmedia>=0.8.0
wagtail-markdown>=0.7.0
gunicorn>=20.1.0
